corrected-migration-file

+++ db/migrate/20231212172527_drop_departments_copy.rb
@@ -1,5 +1,5 @@
 class DropDepartmentsCopy < ActiveRecord::Migration[6.1]
   def change
-    drop_table :departments_copy
+      drop_table :departments_copy if table_exists?(:departments_copy)
   end
 end
